refactor(core): move provider builder out of TUI
build_provider() maps a ResolvedConfig to Arc<dyn CompletionProvider>.
It's infrastructure/factory logic, not presentation. Moving it to
inference::mod.rs removes the TUI's knowledge of how to construct
providers (LmStudioProvider, OpenRouterProvider).
