+++ src/inference/mod.rs
@@ -3,9 +3,31 @@ pub mod provider;
 pub mod providers;
 pub mod types;
 
+use std::sync::Arc;
+
+use crate::core::config::ResolvedConfig;
+
 pub use provider::{CompletionProvider, CompletionRequest, ProviderError};
 pub use providers::{LmStudioProvider, OpenRouterProvider};
 pub use types::{
     Context, ContextItem, ContextSegment, Effort, Source, StreamChunk, ToolCall, ToolDefinition,
     ToolResult, UsageStats,
 };
+
+/// Build a provider from a resolved config's provider name and credentials.
+pub fn build_provider(config: &ResolvedConfig) -> Arc<dyn CompletionProvider> {
+    match config.provider.as_str() {
+        "lmstudio" => Arc::new(LmStudioProvider::new(config.lmstudio_base_url.clone())),
+        _ => {
+            // Default to openrouter
+            let api_key = config
+                .openrouter_api_key
+                .clone()
+                .expect("OpenRouter API key must be set (config file, OPENROUTER_API_KEY env var, or --provider lmstudio)");
+            Arc::new(OpenRouterProvider::new(
+                api_key,
+                Some(config.openrouter_base_url.clone()),
+            ))
+        }
+    }
+}
